+++ examples/udp_client.py
@@ -4,27 +4,27 @@ import iperf3
 
 client = iperf3.Client()
 client.duration = 1
-client.server_hostname = '127.0.0.1'
+client.server_hostname = "127.0.0.1"
 client.port = 5201
-client.protocol = 'udp'
+client.protocol = "udp"
 
 
-print('Connecting to {0}:{1}'.format(client.server_hostname, client.port))
+print(f"Connecting to {client.server_hostname}:{client.port}")
 result = client.run()
 
 if result.error:
     print(result.error)
 else:
-    print('')
-    print('Test completed:')
-    print('  started at         {0}'.format(result.time))
-    print('  bytes transmitted  {0}'.format(result.bytes))
-    print('  jitter (ms)        {0}'.format(result.jitter_ms))
-    print('  avg cpu load       {0}%\n'.format(result.local_cpu_total))
+    print("")
+    print("Test completed:")
+    print(f"  started at         {result.time}")
+    print(f"  bytes transmitted  {result.bytes}")
+    print(f"  jitter (ms)        {result.jitter_ms}")
+    print(f"  avg cpu load       {result.local_cpu_total}%\n")
 
-    print('Average transmitted data in all sorts of networky formats:')
-    print('  bits per second      (bps)   {0}'.format(result.bps))
-    print('  Kilobits per second  (kbps)  {0}'.format(result.kbps))
-    print('  Megabits per second  (Mbps)  {0}'.format(result.Mbps))
-    print('  KiloBytes per second (kB/s)  {0}'.format(result.kB_s))
-    print('  MegaBytes per second (MB/s)  {0}'.format(result.MB_s))
+    print("Average transmitted data in all sorts of networky formats:")
+    print(f"  bits per second      (bps)   {result.bps}")
+    print(f"  Kilobits per second  (kbps)  {result.kbps}")
+    print(f"  Megabits per second  (Mbps)  {result.Mbps}")
+    print(f"  KiloBytes per second (kB/s)  {result.kB_s}")
+    print(f"  MegaBytes per second (MB/s)  {result.MB_s}")